update postAuth comments
